Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    float num;
+    double num;
     printf("");
-    scanf("%f",&num);
-    printf("You entered: %.4f\n",num);
+    scanf("%lf",&num);
+    printf("You entered: %lf\n",num);
     return 0;
 }
